MINOR: mux-h1: Set close mode on client connections if CONNECT method used
Explicitly set WANT_CLO flag on H1 stream, on the client side, if a CONNECT
request is received. This way, an explicite "Connection: close" header will
be added to the response. If the tunnel is established, it doesn't
matter. If the server refuse the tunnel, this only makes explicite what is
done anyway. And it is probably safer to close in such case.